test(builder): extract getStateBuilders response fixture
