+++ packages/builder/test/unit/utils/apiStub.ts
@@ -1,7 +1,7 @@
 import {Mocked, vi} from "vitest";
 import {ApiClientMethods, ApiResponse, Endpoint, Endpoints, HttpStatusCode, IHttpClient} from "@lodestar/api";
 
-type ApiClientStub = {[K in keyof Endpoints]: Mocked<ApiClientMethods<Endpoints[K]>>} & {
+export type ApiClientStub = {[K in keyof Endpoints]: Mocked<ApiClientMethods<Endpoints[K]>>} & {
   httpClient: Mocked<IHttpClient>;
 };
